feat(`monitoringp`): implement `BeginBlock` and report signatures (#486)
* change last blokc param type

* initialize signature reporting

* initialize report tests

* complete report test

* implement transmit signature

* lint

* convert cons address to byte array

* add transmit packet

* fix test

* fix monitoring packet

* fix monitoring packet

* fix nano timestamp

* create monitoring history map

* add monitoring history to recv

* disable channel ordering

* set back ordering

* fix capability

* disable version

* json marshal packet

* refactor monitoring packet

* packet json marshaling

* lint

* add comment for existing issues

* Apply suggestions from code review


* fix error

* fix tests

* merge queries

+++ pkg/types/monitoring_packet.go
@@ -11,7 +11,7 @@ func (m MonitoringPacket) ValidateBasic() error {
 		return errors.Wrap(err, "invalid signature counts")
 	}
 
-	if m.BlockHeight < m.SignatureCounts.BlockCount {
+	if uint64(m.BlockHeight) < m.SignatureCounts.BlockCount {
 		return fmt.Errorf(
 			"block height %d must be greater or equal to block count %d",
 			m.BlockHeight,
